Revert "Fix KnownSpell management: Blizzard did not updated the list of spell managed by the LUA function yet."
This reverts commit 5b62d748c237497737ab1df25ea98e1a9a30ad0c.

We are stuck to wait for blizzard update. Adding this "fix" just make the hunter find the wrong spec. (Survival because the Explosive Shot is found in the spell book even if not "known") without this fix, it found Beast Master, (the default spec), so it's always better.
Adding an ID everywhere wont fix this, explained by apoc here: http://www.ownedcore.com/forums/world-of-warcraft/world-of-warcraft-bots-programs/wow-memory-editing/373699-wow-5-0-4-16016-x86-info-dump-thread-7.html
A probable fix is available the page after, but it's kind of wierd.