Add method SummonedBySpellName in Spell() class.
Note: This is not a perfect function, there is no way to detect if a totem is summoned by your GUID so we both check (Object Guid == Me.Guid), OR (faction==faction, level==level).
We can maybe check the totem title, I guess the name of the caster is inside. (Totems are NPC)