Modify ProfileType structure to allow having advanced data such as : Timer1,2,3 Bool1,2,3 String1,2,3 Int1,2,3.
Check if we changed Battleground in order to reset Profile information.
Implement a working AfkSomewhere ProfileType for Warsong Gulch.
Modify Battleground Resurrection to a more basic way of working, as long as we don't have all the Spirit Healer Ids.
TODO: Make a function that check Ids of Spirit Healer for each faction.