First version of CTF is ready !
Fix Honor management. (When you start the bot, for some reasons, it take some time to get the real amount of Honor (set to 0 instead)) Note: This wont count the first Honor earned if the personn was really at 0 honors. (We don't mind about loosing 2 or 3 honors points in the stats, do we ?)