To be reviewed.
Add Traveler's Tundra Mammoth support.
Add Grand Expedition Yak support.
Plus: Maybe fix ToTown system (were doing an useless addition of "if the same NPC can sell & repair" while we just have to add each npc to the list and it will parse every action), the main goal of that is to avoid going to 2 differents area for the same things, but it get the nearest everytime, so it will go to the same mostly