Regarding what Shauren (tc2) told me, there is no way how getting the real immunity status of a Unit via client memory.
( <%Shauren> mechanic immune is made by hand and is not sent to client )
However, we can check if the target is a WorldBoss, they are immune to stun in ~95% of the case, "some" Boss from dungeon are not immune (source wow wiki).
Some elite are immune as well (source wow wiki), but we wont taking care of that, the "IsStunnable" is more designed in case of fight with the Damage Dealer against a Raid boss.