Add Hapiness to PowerType.
Add some comments to PowerType.
Note: There is only 15 PowerType in FrameXML\Constants.lua, see http://www.wowpedia.org/PowerType
ArcaneCharges does not appear on Constants.lua, we need to test it.
Note: LightForce seems to have been replaced by Chi, and DarkForce seems to be useless for now but not removed, we don't know if it's a Alpha PowerType of the monk, or if it's a placeholder for the future. (The alpha powertype is currently the best guess, as "Light/Dark force" would have been a good Yin & Yang combo, but they seems to have preferred "Chi" ("Qi" - Energy/Vital Power) over the Yin & Yang model (opposite forces))