Fix Main.loop in Healer Class.
If we don't have any target, then automatically target ourself.
If we are targeting an unfriendly unit, then automatically target ourself.
If lowestHpPlayer is not valid, then automatically target ourself.
HardFix ShadowFiend usage, select a hostile creature in combat with me and use it on the creature. TODO: Create a function to select a hostile target for that kind of spell such as Regen spells offensives.
Fix InnerFire/InnerWill spamming, check if the buff is on ourself, not on target as it's a personnal buff.