Allow LongMove to be able to go by foot to a destination if MountCapacity.Feet.
Regenerate the whole path every 1 minutes in case the bot is closer and find a better way.
Not working 100% well, as the bot feel he can pass over mountains then finally arrive to the end of the generated path, then see that he cannot generate a new path since he is in front of a big mountain...
My case (interesting), I were in Vale of Eternal Blossoms, then the bot wanted to go "straigh" way to my profile in Valley of the Four Winds. So he choose the "quick" way. If he wanted to generate the real path, it would be to go to north to Kun-lai Summit by the great door, then go east to The Jade Forest, the go West to Valley of the Four Winds. Arround 3000yards ?