Archaeologist now support 100% ground movements. (Optimist.. It actually stop doing anything if he found 4 uncomplete paths)
Note: The method used here should look a little "not well optimized" since it generate 4 heavy path everytime, but keep cool, it wont affect the normal Flying Archaeologist.
It actually check for every available DigSite, and check the better of them by PathFinder returned Status > Priority > by Distance, so it may load up to 4 PathFinding process for 4 very big distances.
Note: The priority over PathFinderReturnedStatus is really important, if the path is not found to your destination, you will simply end with a long road for nothing, stuck in front of a big mountain separater for 2 zone, etc..