UpdateLastHardwareAction timestamp everytime we CTM or Interact or LuaDoString.
Note: If you are already AFK and the bot natural movements don't get you out of it, you wont get out of AFK by this. Sometimes, the bot is unable to get you out of AFK, or, not enough faster (BG instant kick, Fisherbot without profile never move, so never use the "jump" action).
The fact we don't use PressKey anymore for most of the action also play an important role. (but with my test, an Ascend Start does get you out of AFK)