Fix "Not repairing" issues for ever + explanation.
The GetDurability function were using the VisibleItems Descriptors to get the equipped Id.
The problem is that the VisibleItems Descriptors don't work if the item is Broken (because not shown), and show the Transmogrification instead.
The result of this is that, we were not parsing the items because the item id were not found. (we check if the item is equipped by its id see the "var itemId = new List<uint>" in WoWPlayer.cs)
So, without a valid Id, we were giving up, each Broken items were not returning a "MaxDurability", so, everytime you broke an item, it was not anymore in the count for your % of broken items.
Additionnaly, if you broke ALL items, the MaxDurability now being "0", the percentage generation fail and go in the Catch, and we were returning 100% as a default value, without even outputting a log of what we just did.
I removed the try/catch, I fixed the item id finding, so now all items return a MaxDurability, so we will not have a Division by Zero issue.
I also removed the check that does "if (Durability > 0 && MaxDurability > 0) { we add it in the count }, because if we fixed the item id finding, we would end up with a valid MaxDurability, but because Durability were not > 0, we don't save them in the count for %.