Implement a new AntiAFK system based on the same principle of incrementing LastHardwareAction, but letting WoW doing it, as it does it better than me (Risk Free ! :))*
We don't know which Timestamp WoW really use, we know which function generate it, but we have no reasons to reverse something, when it's not necessary
I debugged the offset LastHardwareAction in Cheat Engine, and it gave me access to the functions that write it, basically: (DownKey, DownMouseButton, UpKey, UpMouseButton, Move with mouse, Move with mouse while on DownMouseButton), there is maybe more action, but I stopped at "Down/Up key".
I'm searching for a Free key with a simple binding (1 key), even if the user has binded all his keys, we have our backups F13 to F24.
I KeyPress it with a reasonable delay of 10millisec between, I call the whole function every 0.5 seconds, I keep the lock, because I were having 2 simultaneous task here (probably because of a loop I caused before editing LuaDoString to not use my function UpdateLastHardwareAction if called from this function.
Enjoy :) I have been testing with Cheat Engine, while fishing etc, it will correctly update the LastHardwareAction, and this time, it will even get out of AFK, as we don't just erase the value in the memory.
Remove management of the offset LastHardwareAction.