Advise to keep Do Avoid Melee OFF + set range to 0 (deactivated). This is causing 'Too Much' issues.
Todo: Move the function out of CombatClass, rewrite it, add a timer or a count for each combat to avoid spamming it.
If not done, it will be simply removed as it serve no purpose, even in the kind of combat where it should be needed, it don't do its work.