US Computer comes with Accessibility parameters on font size 125% per default, maybe it's because they are fat and blind, but the software have been developped for a 100% in settings.
I understand why some people says "Shit bots" now, the fix of today will allow the form to scale with the font size scalling.
Note: I set it to 100F rather than anything else after having made some tests, the Profile Creator of the Grinder only show the "Record" button with 100F +, but if you set it too high, it make the form too big, so it's the perfect parameter to me. (normally, people recommand setting it to 96F, because of the 96DPI for a 100% user (120 for 125))