Support 100% of the Blood Elf Starting Zone. From Warrior (1) to Monk (512).
Rename the profiles with a better naming style.
Blood Elf Starting Zone support only MaskRace for Blood Elf (as well as Draenei Starting Zone for Draenei).
The quest at the door of the starting zone allow all the Horde though. But the quest that lead to that NPC is still Blood Elf only quest.