Add some quests into Eversong Woods profile.
There is 2 quest where they ask you to take the flypath, but it's not in the quest objective so we can go by walk.
There is 1 quest (Whitebark's Memory) that will require some more coding, the objective is SelfCompleted, so the bot is trying to TurnIn first, but in fact, you need to Use an Item, Fight the NPC, talk to the NPC, read all the conversation (page or gossip, I haven't completed it to be able to test later), then you finally need to turn in the NPC to the monster.
There is still arround 10 quests to add, I have been a bit busy and tired today...