Fix products loading (unproper way).
I first wanted to properly define the namespace, but haven't taken care of the fact the bot use "Main" to direct load (sorry).
So this is a temp-fix (simple revert) and will be followed by a complete fix for loading the proper namespace for each DLL when I will have the time.
basically: Damage Dealer => Damage_Dealer, we just have to replace space with an underscore and load as is.
Damage_Dealer.Main instead of Main.
this is not so important so I will do this when I have extra time as this may cause issue with people having their own CC not using a namespace at all for example. so why not first loading Main alone, and if null, trying with a namespace for example..
I will think about it, currently, keeping "Main" is the easiest way.