Implement and use IsNewSkillAvailable(int value, int maxValue, SkillLine skillLine, bool hardCheck = false).
Note: Price and Levels of skills are hard-coded. But verified ingame with Friendly reputation to Orgrimmar.
Todo: if useful, find the next spell cost/level required from the game (but we will need to check the basic price neutral (usually friendly with capitals) since we don't know yet which NPC we will have)