First part of TrainNewSkill system done.
Todo: Add a check for every skills in the list of skills to add. (we don't have any NPC for them yet)
Todo: Make the "fakeSettings" a real global settings var.

For the second part:
We first need an algorytm to order NPC by the shortest distance between each over, so we may take the less time to finish the loop. (imagine a circle in a village, rather than talking to Left NPC first, then right, then left), you just talk to each one by one from left to right. Now imagine it in a 1000yards radius range.