Revert Awesomness. Explanation below:
First of all, to debug Tracker etc, we need to replace the .Show by a ShowDialog.
Also, to fix an issue with the little windows like Gatherer profiles selection, we need to always reset the variable of the thread,
if it used (if (var != null && var.IsAlive) var.Abort()). (that was a great patch I did on my testing) Note: I also needed to Abort StopThread in Start and vice versa. So it was updating the button properly. Another trick was to keep the thread open while doing this (for tracker etc), but it's freeze the tracker windows, that's why I preffered ShowDialog.

Now, the worst issue is that, from a thread to another, it's very hard to be able to create dynamic forms like "OpenFileDialog" without freezing the whole application.
And many more things to do. If we want to separate products from the mainframe, we must do it good.

Side Note: Even if you don't understand what I'm talking about, it's a good history info for me, I will.

We are back to 5sec freeze on loading of products :/