2 Fix for AutoSkillTraining empty's slot detection.
Fix "NPC Found" by adding an extended check if we are not moving but not yet arrived.
Mining and Herbalism are not using the spell name "Mining/Herbalism" in the spellbook, so we use Smelting for mining and Herb Gathering for herbalism.
Add LaunchOnSelf with options and casttime.
Make LaunchOnSelf() directly go though Launch with parameters.