Make sure we check baseAddress after all FindTarget calls.
Add a place-holder into QuestingTask to detect and remove a wrongly placed QuestGiver.
Modify ToTown to handle the logging only if the action have been done when possible to check. (not possible for the selling items one)
Allow FindTarget to be passed with parameter doMount = false so it wont dismount/mount ToTown when on Mammoth or Yak.
Minor design change (baseAddress != 0 ==> baseAddress > 0).
Do not closeMerchant after selling/buying/repairing until we reach the end of the Npc loop.
Allow to DoMilling/Prospecting InTown if the current NPC is not repair/sell and magic mounts are ON. (will work with mailbox then)