Update/Improve CheckInventoryForLatestLoot(int eventFireCount).
Do not check the Guid of items.
Now set count to 0 if an item is completly missing so we wont have to completly add it back when we loot it again. (example: motes with AutoMakeElemental)
Change a LINQ expression to normal foreach in ObjectManager.cs.
Use method Owner in DevTool.