Relocate Communication.cs to nManager.Helpful to allow access from anywhere to its methods.
Add Communication.Shutdown().
Listen only if the settings was set to true at the starting of the bot.
Automatically Listen() or Shutdown() when updating the Mimesis Broadcasting settings if the bool changed.
Now display in the normal log when Listen/Shutdown action happened.
Change field type from NumericUpDown to IntegerInput.