Implement Auto grouping the clients bots Also changed the bytes sent to master/client to include data len if any, without this, the server/client read several packets at once. Then no more "new byte[4096]" which wasts memory.
Need to test the grouping code with a client+master on the same server since I added -realmName to the char name to be able to invite cross-realms.

PickUp quest works but TurnIn does not since QUEST_FINISHED seams to be not the proper event and it is fired up for an unknown raison on quest accept.

Signed-off-by: Neo2003 <Neo.2003@Hotmail.fr>