After more check, this sleep here seems to be 100% useless, thanks to blizzard for giving us Filler spell in our DPS cycle.
So this Sleep is called 1-5 times between a spell... "1-5 fois de trop" to me. I noticed I still have a very slight delay of 0.1~ seconds before casting the next spell when my GCD ends, this will reduce it a bit even if looks useless.
The QuickCast function accoupled with a custom Launch() should fix that. (that custom launch should spam the spell until it goes on cooldown for real)