CombatClass: Revert sleep removal Let's start to explain why I revert this.
Basically, without sleep the combat class is doing while(true) { read lots of stuff in wow memory }
So this thread is using 100% of 1 cpu core and is doing millions of queries slowing spellmanager and objectmanager such a way that this thread is often not responding for several seconds

Anyway this sleep does not slow down the combat class a lot, it makes it work.
Imagine we are in combat, then we are not on a mount, inFight, with a target and in range
The code then is:

while (Main.InternalLoop)
{
    Combat();
    Sleep(100);
}
We can loose from 0ms to 100ms depending on when in the GCD the code is called.
Then we loose around 50ms per GCD, 1/20 of a second, it's not that much.

And there were some badly placed Sleep(500).

Signed-off-by: Neo2003 <Neo.2003@Hotmail.fr>