Don't go back to previous point when farming :) Finaly done.
This code will select a point between the last one and the 10th ahead.
It then use 2 vectors defining the directions (me->point) and (point->point+1) and check the 2D angle.
If the angle is < Pi/3 (60°), then it means we won't do a turn-around, so select the point.
If we don't find a suitable point, then we do like before and go back to last point.

Signed-off-by: Neo2003 <Neo.2003@Hotmail.fr>