Bring more offsets to the bot.
Add few commented stuff to find badass offsets.
Fix Injections by removing call to "ClntObjMgrGetActivePlayer" as because of the new 128 bits GUID it's kinda fucked.
Add a few DBC offset manually (function to retrieve them is fucked), structure seems to be changed and/or the way of reading is finally fuck'd up for all of them)
Removed DBC we don't read anymore for a while.