Update TnbAutoInterrupt to 1.3.0.
Improve TnbAutoInterrupt plugin, now works in Duel/Arena for checking ennemy player of the same faction.
Check if the CastingSpell is still valid when we found our interrupter spell or sometimes it returns "kicked spell 0" because the damage dealer kicked it first.
Check for IsCast while checking if a spell in interruptible.