Update BuffManager, implement class UnitAura(s).
Example usage for refreshing dots on a target :

            Spell myDotSpell = new Spell("myDotSpell "); // declared outter scope in CombatClass
            var testAura = ObjectManager.Target.UnitAura(myDotSpell .Ids, ObjectManager.Me.Guid);
            if (!testAura.IsValid || testAura.AuraTimeLeftInMs < 1500)
                myDotSpell .Launch();
Will refresh my dot if the target don't have this dot (from me) or the dot is going to expire in 1.5 seconds.

It's basically a "HaveBuff" but a bit more complicated.

Also, example of usage for refreshing a self buff on self :
            Spell WaterWalking= new Spell("Water Walking"); // declared outter scope in CombatClass
            var WaterWalkingAuraInfo = ObjectManager.Target.UnitAura(WaterWalking.Ids);
            if (!WaterWalkingAuraInfo .IsValid || WaterWalkingAuraInfo .AuraTimeLeftInMs < 30000)
                WaterWalking.Launch();
Will refresh WaterWalking if I don't have it or if it's going to expire in less than 30 seconds.