Improvement in Hooking part.
Will now try to automatically rehook only when the bot has successfully hooked the game and is not going to shutdown.
Disable ReHooking when starting the bot / shutdown'ing the bot, Active it after first hook.
Move UnHooking code to DisposeHooking, if it failed, don't force a "try-hook" and notify the user instead, it wont works anyway and will probably crash WoW.
Those rare case where the bot wouldn't be able to unhook are rare and we need those users to force come to us with informations so we can handle their system.