Implement Druid Travel Form.
Drop support for Druid old forms.
Enable Druid to farm under water nodes back.

(Code in MountFlyingMount go out of water in order to cast a flight form. the new travel forms will automatically convert to a flying form when jumping out of water, so I disabled the rest of the code if you are actually on a druid mount)