Fix Travel path generation when intercontinental. (was calculating the wrong path)
Add a way to go back to departure quay if failed to stay on the quay. (example, fall when trying to enter a zeppelin).
ToDo: finds a way to enter a Zep/Boat 100% chance since their center point is often way off the Quay pass. (cannot either "go in front" as we would need to save our rotation as well when doing the "Outside" point.