Add (information) logs to Travel.
Fix "waiting to travel" by doing a StopMove before actually jumping. (jump in order to validate the TransportGUID in an Elevator)
If you save a good orientation, you can travel using a Zepplin ATM, but you must be a bit on the left/right of the quay, the perfect solution would be to calculate weither the Zep's noze is facing your left or your right and then adjust your destination regarding its size in % of his size. So it wont break elevators for example.