Update Sleeping in QuestingBot.
(the bot is actually spending approximatically 5 minutes AFK on 60 minutes because of too many sleeping.)
This is a test and if there is no issues, it will be good to go.

Also revert "gives some time to refresh ObjectList", ObjectList have no problem at all to refresh, it was simply because the unit were blacklisted, fixed since I added "IgnoreBlackList".