[TEST] Try remount 6 seconds after UnStuck dismounted us because it never ever going to remount us until the next farm. Can be a 1minute more on feet.
Please report any issues or feel free to edit if you feel this is wrongly placed.
The purpose of that is to automatically re-mount if not mounted but can mount while moving with MoveTo() (core function for all movements).
As you can see, there is a minimum distance to remount of 12 yards to avoid remounting when we wanna move only few yards away (archaoelogist ? :)
Also wont try to remount if in fight or dead. Etc...