Try some stuff in Test_Product to be able to automatically finds/adds new taxi nodes.
Goal is not just to easily parse draenor but the whole World of Warcraft.
Added a field in Taxi class to be able to identify an unknown taxi by its XY when not knowing it.
Add GetNumRoutes parameter to see what taxi are only 1 hop away from our position.
Also, taxi with 0 hop are either "unknown or opposite faction"... or just the one we are currently talking to !