Allow KillMob objective to have only a single "<Position>" defined when no hotspots.
Get ride of untranslated log and just write something readable in english.
Get ride of "hotspot 0 to hotspot 0" non-sense.
Add a logging when we don't find a hotspot so the users know what's going on if the bot is not doing anything.
Hack ObjectiveCompleted in an InternalQuestId that have been already turn'd In.