Gives lower priority to IsQuestFlaggedCompletedLUA.
Check if the quest have been completed after we tryed to pick it up. (Fixes the bot trying to take the same "follower/self completed quest" over and over until you restart the product).
(the ObjectiveIsComplete check is done before, you never do it again after entering "ExecuteObjective", which end up in a loop of trying to complete an objective that was done :p)