Important : Fix opCode "ReplyPosition" in Mimesis. (see below)
We had an issue with "ObjectToBytes", this function would completly dump the instance of the master Point(), including, very useless data. ie: My Position was 261 bytes long...
The problem with that is that we are currently only using 2 Bytes for sending opCode (coded on 1 byte) + lenght (coded on 1 byte).
I didn't find usefull the possibility of increasing either opCode size or lenght size. 261 bytes for 3 array and a little string is a bit too much, especially if it's going to work on remote computer.
So instead, I added the 3 float "byte to byte" + a 1byte value which represent the Point.Type. (where 0 = None, 1 = Swimming, 2 = Flying).

Now, a ReplyPosition opCode lenght is 13 bytes. (4+4+4+1).