Use GetSpellCharges for Supernova. Remove timer.
Don't return after using IceFloes or it totally break the concept of it.
Don't check main target distance when we already check how many units are in range. (lol)
Turn UseEvocationAtPercentage to UseEvocationBelowPercentage. (since it's a < 50 check, not a <= 50, you can lose DPS doing that, the goal is to go Under 50).

Note: When possible, uses the Spell.Range in GetUnitInSpellRange instead of a hardcoded value.