Fix "StrikeBack" thread freezing the bot going into an endless loop of pathfinding.
I simply ignore UnitInAggroRange while we are mounted, etc, etc as it should be, it no longers stuck into an endless pathfinding.
I haven't found a way to resolve the pathfinder itself, the path requested was not necesserly big, but at least, instead of having the issue every 20minutes, I don't have the issue anymore.