OnSpot fishing bot now can properly returns to position if can no longer fish.
Does not stop at checking if the fishing loop is still running, it doesn't mean at all that we can still fish, instead, save position only if we looted a fish in the last 5 sec, and go back to original point if we did not loot anything in 25secs.
Also, can now equip its weapon back when entering combat, and will return to orignal position after combat ends. (instead of saving the incombat position)
No longer move a bit forward the first time we start it, and we should add extensive distance check with recorded point so we don't move if < than CTM precision.