Reduce MaxNodes to 262143 as suggested by @Neo2003.
Also increase maxTiles up to 4096 as some continents really have alot of them and the pathfinder currently crash if we don't allow all of them to load up in very very big Travel + already loaded data of a region in Quester.
Update MaxPolygons to 20bits, I'm not sure about that, I've just figured out that TrinityCore always uses 1 << polyBits and polyBits is set to 20 per default on 64bit defined version of Detour.